created a next button component
