+++ src/navigation/MainNavigator.tsx
@@ -10,7 +10,7 @@ import {PrivacyScreen} from 'screens/privacy';
 import {OnboardingScreen} from 'screens/onboarding';
 import {LanguageScreen} from 'screens/language';
 import {useStorage} from 'services/StorageService';
-import {RegionPickerSettingsScreen} from 'screens/regionPicker';
+import {RegionPickerSettingsScreen, RegionPickerScreen} from 'screens/regionPicker';
 import {NoCodeScreen} from 'screens/nocode/NoCode';
 import {HowToIsolate} from 'screens/howToIsolate/HowToIsolate';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
@@ -72,23 +72,22 @@ const NoCodeWithNavBar = withDarkNav(NoCodeScreen);
 const HowToIsolateWithNavBar = withDarkNav(HowToIsolate);
 
 const OnboardingStartWithNavBar = withDarkNav(Start);
-const OnboardingPermissionsWithNavBar = withDarkNav(Permissions);
+const OnboardingWhatItsNotWithNavBar = withDarkNav(WhatItsNot);
 const OnboardingAnonymousWithNavBar = withDarkNav(Anonymous);
 const OnboardingHowItWorksWithNavBar = withDarkNav(HowItWorks);
-const OnboardingWhatItsNotWithNavBar = withDarkNav(WhatItsNot);
+const OnboardingPermissionsWithNavBar = withDarkNav(Permissions);
+const OnboardingRegionWithNavBar = withDarkNav(RegionPickerScreen);
 
 const OnboardingStack = createNativeStackNavigator();
 const OnboardingNavigator = () => {
   return (
-    <OnboardingStack.Navigator
-      screenOptions={{stackAnimation: 'fade', headerShown: false}}
-      initialRouteName="OnboardingTutorial"
-    >
+    <OnboardingStack.Navigator screenOptions={{headerShown: false}} initialRouteName="OnboardingStart">
       <OnboardingStack.Screen name="OnboardingStart" component={OnboardingStartWithNavBar} />
-      <OnboardingStack.Screen name="OnboardingPermissions" component={OnboardingPermissionsWithNavBar} />
+      <OnboardingStack.Screen name="OnboardingWhatItsNot" component={OnboardingWhatItsNotWithNavBar} />
       <OnboardingStack.Screen name="OnboardingAnonymous" component={OnboardingAnonymousWithNavBar} />
       <OnboardingStack.Screen name="OnboardingHowItWorks" component={OnboardingHowItWorksWithNavBar} />
-      <OnboardingStack.Screen name="OnboardingWhatItsNot" component={OnboardingWhatItsNotWithNavBar} />
+      <OnboardingStack.Screen name="OnboardingPermissions" component={OnboardingPermissionsWithNavBar} />
+      <OnboardingStack.Screen name="OnboardingRegion" component={OnboardingRegionWithNavBar} />
     </OnboardingStack.Navigator>
   );
 };
