+++ src/screens/onboarding/components/BackButton.tsx
@@ -1,0 +1,25 @@
+import React, {useCallback} from 'react';
+import {useNavigation} from '@react-navigation/native';
+import {useI18n} from '@shopify/react-i18n';
+import {Button, Box} from 'components';
+
+export const BackButton = () => {
+  const [i18n] = useI18n();
+  const navigation = useNavigation();
+  const prevItem = useCallback(() => {
+    navigation.goBack();
+  }, [navigation]);
+  return (
+    <Box flexDirection="row">
+      <Box align-self="flex-start">
+        <Button
+          backButton
+          text={i18n.translate('Onboarding.ActionBack')}
+          color="linkText"
+          variant="subduedText"
+          onPress={prevItem}
+        />
+      </Box>
+    </Box>
+  );
+};
