+++ src/screens/onboarding/views/WhatItsNot.tsx
@@ -3,8 +3,10 @@ import {ScrollView, StyleSheet} from 'react-native';
 import {Box, Button, Text, OnboardingHeader} from 'components';
 import {useI18n} from '@shopify/react-i18n';
 import {useNavigation} from '@react-navigation/native';
+import {SafeAreaView} from 'react-native-safe-area-context';
 
-import {BaseOnboardingView} from '../components/BaseOnboardingView';
+import {BackButton} from '../components/BackButton';
+import {NextButton} from '../components/NextButton';
 
 export const WhatItsNot = () => {
   const [i18n] = useI18n();
@@ -12,30 +14,35 @@ export const WhatItsNot = () => {
   const onNext = useCallback(() => navigation.navigate('OnboardingAnonymous'), [navigation]);
 
   return (
-    <BaseOnboardingView>
-      <ScrollView showsVerticalScrollIndicator={false} contentContainerStyle={styles.content}>
-        <Box flex={1} paddingHorizontal="xl">
-          <OnboardingHeader
-            text={i18n.translate('Onboarding.WhatItsNot.Title')}
-            imageSrc={require('assets/onboarding-neighbourhood.png')}
-            accessible
-            accessibilityLabel={i18n.translate('Onboarding.WhatItsNot.ImageAltText')}
-          />
+    <Box flex={1} backgroundColor="overlayBackground">
+      <SafeAreaView style={styles.flex}>
+        <BackButton />
+        <Box flex={1} paddingTop="s" justifyContent="center">
+          <ScrollView showsVerticalScrollIndicator={false} contentContainerStyle={styles.content}>
+            <Box flex={1} paddingHorizontal="xl">
+              <OnboardingHeader
+                text={i18n.translate('Onboarding.WhatItsNot.Title')}
+                imageSrc={require('assets/onboarding-neighbourhood.png')}
+                accessible
+                accessibilityLabel={i18n.translate('Onboarding.WhatItsNot.ImageAltText')}
+              />
 
-          <Box marginBottom="m">
-            <Text variant="bodyText" color="overlayBodyText">
-              {i18n.translate('Onboarding.WhatItsNot.Body1')}
-            </Text>
-          </Box>
-          <Box marginBottom="m">
-            <Text variant="bodyText" color="overlayBodyText">
-              {i18n.translate('Onboarding.WhatItsNot.Body2')}
-            </Text>
-          </Box>
+              <Box marginBottom="m">
+                <Text variant="bodyText" color="overlayBodyText">
+                  {i18n.translate('Onboarding.WhatItsNot.Body1')}
+                </Text>
+              </Box>
+              <Box marginBottom="m">
+                <Text variant="bodyText" color="overlayBodyText">
+                  {i18n.translate('Onboarding.WhatItsNot.Body2')}
+                </Text>
+              </Box>
+            </Box>
+          </ScrollView>
+          <NextButton onNext={onNext} />
         </Box>
-      </ScrollView>
-      <Button text={i18n.translate('Onboarding.ActionNext')} variant="thinFlat" onPress={onNext} />
-    </BaseOnboardingView>
+      </SafeAreaView>
+    </Box>
   );
 };
 
@@ -43,5 +50,8 @@ const styles = StyleSheet.create({
   content: {
     flexGrow: 1,
     justifyContent: 'center',
+  },
+  flex: {
+    flex: 1,
   },
 });
